Add files via upload
This project comes under Frontend and Backend Development.
 The required tools are 
1) Frontend -- HTML,CSS and Bootstrap 
2) Backend -- Python Programming  

Description:- The Telephone Directory consists of three major Functions. there are Create, Display and Search  For storing the data I used Dictionary. As we know that the dictionary stores the data in Key and Value pairs. Here keys are nothing but Names and Values are nothing but Numbers  

1 - Create Function:- The name and number should be created by the user. And those names and numbers should be stored in a dictionary  

2 - Display Function:- It should represent the names and numbers which the user had created in Create function in a tabular format  

3 - Search Function:- It should represent the data present in the display function. First, it will ask the name to check whether the name exists in the dictionary or not. If the name is present in the dictionary then it shows the name and number otherwise it throws an error like Name not exists